Add retained TUI view engine with clipping and typed event routing.
Resolve ancestor clips and composable geometry in fpas-std, route keyboard/pointer events through capture/target/bubble phases, clip console paint to view bounds, and document the new host model with tests and examples.

+++ crates/fpas-vm/src/tests/core/tui_host_vm/commands.rs
@@ -258,8 +258,8 @@ fn tui_host_view_command_shortcut_uses_focused_ancestor_binding() {
         loc(),
     );
     chunk.emit(Op::GetLocal(0), loc());
-    emit_constant(&mut chunk, Value::Integer(1));
-    emit_constant(&mut chunk, Value::Integer(1));
+    emit_constant(&mut chunk, Value::Integer(11));
+    emit_constant(&mut chunk, Value::Integer(11));
     emit_constant(&mut chunk, Value::Integer(1));
     emit_constant(&mut chunk, Value::Integer(1));
     chunk.emit(
